refactor /tags
