+++ src/features/tags/consts.ts
@@ -1,0 +1,1 @@
+export const COLORS = ["Blue", "Green", "Red", "Purple", "Orange", "Yellow", "LightBlue"];
